server/requirements.txt
mcp[cli]>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pyodbc>=5.0.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
python-dotenv>=1.0.0
azure-identity>=1.15.0
